+++ metasrv/src/meta_service/meta_service_impl.rs
@@ -22,10 +22,10 @@ use common_tracing::tracing;
 
 use crate::meta_service::GetReply;
 use crate::meta_service::GetReq;
-use crate::meta_service::LogEntry;
 use crate::meta_service::MetaNode;
 use crate::meta_service::MetaService;
 use crate::meta_service::RaftMes;
+use crate::raft::types::LogEntry;
 
 pub struct MetaServiceImpl {
     pub meta_node: Arc<MetaNode>,
