+++ metasrv/src/raft/raft_impl.rs
@@ -1,4 +1,4 @@
-// Copyright 2020 Datafuse Labs.
+// Copyright 2021 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use async_raft::LogId;
-pub use async_raft::NodeId;
+use async_raft::AppData;
 
-pub type LogIndex = u64;
-pub type Term = u64;
+use crate::raft::types::LogEntry;
+
+impl AppData for LogEntry {}
